refactor: Cleanly transition mapping parser logic to the anibridge.utils package
