+++ pyproject.toml
@@ -116,7 +116,8 @@ module-name = "anibridge"
 namespace = true
 
 [tool.uv.sources]
-py-spy = { git = "https://github.com/czardoz/py-spy.git" } # Until upstream suports Python 3.14
+anibridge-utils = { git = "https://github.com/anibridge/anibridge-utils.git" }
+py-spy = { git = "https://github.com/czardoz/py-spy.git" }                     # Until upstream suports Python 3.14
 
 [build-system]
 requires = ["uv_build>=0.10.9,<0.11.0"]
